// src/features/profile/types/user.ts
export interface UserProfileData {
  dateOfBirth: Date; // Correctly typed as a Date object
  gender: string;
  bloodGroup: string;
  nationality: string;
  email: string;
  mobile: string;
  alternateMobile: string;
  fatherName: string;
  fatherOccupation: string;
  motherName: string;
  motherOccupation: string;
  guardianName: string;
  fullAddress: string;
}